feat(myji): add support for retrieving API token from system password store
This change allows `myji` to retrieve the JIRA API token from the system
password store using the `pass` utility if it's available, and the token
is not provided via environment variable or the constructor parameter.

This improves security by allowing to use an external tool to manage
secrets.
It uses the environment variable `JIRA_PASS_TOKEN_KEY` or `jira/token`
if none is set.

The function `get_pass_key` has been implemented in `utils.py`.