feat: Open ticket in browser with itself
Adds functionality to open a selected Jira ticket directly in the
browser from the fzf preview window.

The `enter` key binding in fzf is changed to execute the
`myj fzf browser-open` command, which takes the selected ticket
number as an argument.  The `browser_open_ticket` function in
`utils.py` is used to open the Jira ticket in the user's default web
browser using `webbrowser.open()`. A new subcommand `fzf
browser-open` to handle this functionality.