feat: Add tool to add comments to Jira issues
Added a new tool to the server that allows users to add comments to specific
Jira issues. This functionality was previously unavailable and is now
accessible via the "add-comment" tool.

The server now exposes the "add-comment" tool, which takes a ticket key and
comment text as input. A corresponding handler function, `_handle_add_comment`,
was implemented to call the Jira client's `add_comment` method.

New unit tests were created to verify the correct registration and
functionality of the "add-comment" tool, ensuring that comments are added to
the specified Jira issues.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>