feat: Add direct JQL query option to browse and gencontext
Introduced a `--query` or `-q` option to both the `browse` and `gencontext`
commands, allowing users to specify a JQL query directly. This bypasses the
need to select a pre-configured board and provides greater flexibility for
searching and context generation.

The `browse` command now uses the provided JQL as the primary search criteria.
If `browse` is called with both a board name and the `--query` option, the
board name will be appended as a search term to the custom JQL.

The `gencontext` command will use the provided JQL if the `--query` option is
present. If no board is specified along with the `--query` option, it defaults
the board name to "Custom Query". If neither a board nor a query is provided,
an error message is displayed.

Added new test files `tests/test_browse_query.py` and
`tests/test_gencontext_query.py` to cover the functionality of the new
`--query` option for both commands.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>